Add “optional” to addEvent to allow error to be passed back to client
Sometimes we try to add an event that has a good chance of failing.
Currently when we repond to KeySolicitations we first post a KeyFulfillment to the channel. Lots of users can post this fulfillment at the same time, but the node will only accept one of them

After this change, instead of throwing tons of errors polluting all logs, we can now pass “optional: true” and check the error in the calling code