Increase test coverage for joinSpace in sdk (#64)
Increase test coverage for joinSpace from sdk tests to include user
entitlement checks and linked wallets. Check for joinStream rejections
on the stream node that throw permissions errors when the user should
not be permitted to join.

NOTE: these tests are not actually exercising entitlement checks on the
stream node for joinSpace. All space joins happen in the contract and
the stream node checks for space membership before allowing a user to
join the space stream, and other "join" checks are for channel
permissions... so a lot of the recent changes to the stream node around
entitlement evaluation are still not exercised. However this does
increase the coverage in general.

I'd like to follow up with some test cases that add a role allowing a
user to Add/Modify channels in order to test this code.