Banning fix plus test cases (#139)
Some nil pointer exceptions showed up on omega because the banning
interface token contract was not properly initialized. This was not
showing up in test cases or on gamma because no one was banned anywhere
entitlements were checked.

This PR fixes the bug and adds a test case to check for this specific
scenario - a banned user should not be able to join a channel, and also
banning checks should work when a user is banned. In addition, I added a
sanity check test case to the simple cache used by the banning interface
for better coverage.