SyncAgent - wire up user memberships
I think this is the pattern i’m going to stick with for now
The sync agent wraps the old client
The views act like useEffects, they register with the client connection and pass back an uninitialization function
This should allow me to easily move code over from the lib while getting persistence for free.
There’s still a lot of work but I think it’s possible to make a lot of progress this week.

lint

fix user initialization status

switch to unentitled tests for now