BaseRegistry helper function to connect base nodes to their operators (#272)
Part of the on-chain metrics requirements is to help with detecting
discrepancies between nodes and operators on Base vs River.

RiverRegistry's INodeRegistry facet's `getAllNodes()` makes it easy to
query the operator of a given node, since `operator` is part of the
`Node` struct. However nodes are expressed as addresses only on Base,
and we're required to dispatch additional queries to connect the node to
its operator.

This PR introduces a new helper function to BaseRegistry. This helper
function streamlines this process of getting all nodes along with their
operators. We can now use its output to dig into the discrepancies
between River and Base.