Change the miniblock_candidate primary key to be covering for the queries we run. Change one fetch to avoid a sort (#751)
I've reviewed all of the queries that were resulting in serialization
errors and I'm unable to figure out why we would be having any errors in
the first place. It's a red flag that our transaction order, which is
quite small, is causing a serialization error.

Along the way I found two queries that we doing some table level access
that should have been index only. This fixes those cases.

---------

Signed-off-by: Brian Meek <brian@hntlabs.com>