bugfix(sdk) Udpate client.waitForStream (#786)
waitForStream was accepting a string or uint8array and was then
comparing to a string. It was never called with a uint8array, but that
was a bug waiting to happen

it also didn’t timeout, so if the stream never initialized, initStream
would hang forever.

add timeout so that the ui can continue and not hang on loading screens.