core: remove lock contention in transaction pool (#761)
The current transaction pool has several tasks:
- submit new transactions
- track submitted transactions for inclusion in the chain
- send a replacement transaction when needed

Because all tasks operate on the pending transactions list they share a
mutex to prevent race issues. This mutex is also used to prevent nonce
collisions. This causes lock contention and can lead to delays when
submitting and retrieving transaction confirmations. This change splits
the logic into a transaction pool that is used to submit transactions
and an internal pool that deals with pending transactions.

The mutex in the transaction pool is only used to prevent nonce
collisions on submit. The pending tx pool uses `sync.Map` that allows
concurrent use and therefore allows for parallel inclusion of newly
submitted transactions and checking if transactions are included or need
to be replacement.

This approach removes lock contention preventing sometimes slower
transactions confirmations and scales better as the number of
transactions increase.

After this change the xchain integration tests failed due to nonce
problems. Those were caused by creating multiple transaction pools on
the same chain for the same wallet and is now fixed.