core/node/crypto: optimize algo that calculates next time to poll for new chain head (#780)
Currently the chain monitor starts up and polls each block period for a
new block. If the monitor has its poll interval just after the moment a
node usually has the block available this is fine but if the monitor
polls at the "end" of the block period this can introduce an almost 2s
delay when receipts/state changes are available.

This change optimises this poll strategy by reducing the next poll
interval until it was too fast and no block was available. If that
happens it will retry it again 2 times after a short wait. If that fails
it waits block period.

The result is that all nodes convergence over time to the moment their
rpc node has a new block available. This reduces the delay after which
state changes are available (tx receipt, events).