Space/Channel entitlements should be granted if a single role's ruleData evaluates positively (#853)
Issue: harmony was requiring user to satisfy multiple roles instead of a
single role to join a channel. This turned out to be an issue in how
wallets are evaluated in the spaceDapp - user entitlements properly
evaluated, but rule entitlements were restricted so that users could
only be granted permission if they satisfied all.

This PR fixes the logical error and adds a new test case that failed
with the previous implementation and passes with these changes.