flatten onStreamInitialized
Previously, all member metadata models has listening to the stream initialized event, causing uneeded subscriptions.

`Membership` model was listening to `streamNewUserJoined`, which is a uneeded listening, since this model lives in the `Member` model, which lives in the `Members` model, which gets the `streamNewUserJoined` too to update the member state.