Add helper functions to generate and parse “auth token”
If you control the signer for a private key that owns a river user you can
- pick an expiration date
- create a new delegate wallet
- sign the new delegate wallets public key and expiration date with your signer
- export the private key, signature, and expiry as a hexified binary representaion of an AuthToken protobuf

Now you can create a signer context anywhere.

This auth token can be used to sign events and add them to river streams, but it can’t be used to make blockchain transactions.

it prints out a string like this:
```
0a4230786637666335656233636636633633376432326461393730363937316639326264313930336335663966643835306166373534353635626166623766366463663612414465bfdd7b9c56ebfcf93bf6e1767ac497d7fc0820589f2d3e3d61415d9e13060c7acd80b991281ce8fc31e0d2553298b20e2eb828abc31a93dcc64812a4f3561c18a6bb808e9a32
```