Update ETH balance go implementation to consider all ether-based chains (#986)
This PR also includes configuration for stream nodes that defaults to
sane values if not explicitly set. It uses the same design as existing
Xchain configuration and should be similarly forward-compatible with the
move to on-chain configuration.

This PR leaves the repo in an inconsistent state, as the client is still
evaluating eth checks on a single chain - however, the eth check is
already broken because it does not meet the spec, so I figured this was
fine.