Fix media entitlement test that mistakenly always created legacy spaces (#1042)
One stray test was not migrated to branch on which space version to
create depending on test suite, and is always making legacy spaces. This
PR fixes that test to use the space determined by environment variables
to E2E suites run locally and in CI can correctly exercise both legacy
and v2 spaces.