set up space factory init for upgrades on diamond cut (#1057)
This PR helps in using a deployed SpaceFactoryInit contract during
diamond cut proposal. Helpful in the event that we need to set state
parameters right after creating the diamond cut during governance. For
example in this case we need to set a SpaceProxyInitializer contract
address right after the cut in order to not break createSpace in
production.