Refactor StreamRegistry.getPaginatedStreams to never return empty streams (#1053)
ref #507 

getPaginatedStreams would return an oversized array in cases where the
stop argument was > # of streams. This PR repairs the contract response
and adds test coverage for the getPaginatedStreams method.

Once this is deployed in all environments, we can remove the special
case code in the stream node that ignores empty stream responses.