Kerem/dd trace with docker builder (#1115)
with our unique combination of yarn workspaces + docker with builder
pattern + esbuild dependency externalizations, our stream-metadata
service's instrumentation was only half-working. spans and traces were
emitted but profiling was not. this PR explicitly reinstalls the
externalized package in the runner stage in order to maintain access.

i'm also setting sourcemaps:true instead of sourcemaps:inline because
i'm not happy with how our other existing instrumented services show up
on the profiler - they are not properly tied to the source code and i'm
just curious to see if this will make a difference.