Encryption keysharing: update initial keysharing timeout
Wait at least 30 seconds to share keys before allowing a user to send messages to a channel

This has always been an optimization to enable users to quickly send messages to a channel, but I think it is no longer correct because
1) we have optimistic message renders that will give the user imediate feedback
2) iOS waits for all keys to be shared before posting a message and the experience is good there

I’ve included a timeout because as long as keys have reached some people in the group extended keysharing will allow the rest of the group to share keys with each other.

The 30 second time is arbitrary, meant to be a comprimise between forever and not long enough.

Future improvements - i think that we could optimize by sending to your own devices and recently active users first, but this is much more work and may not be worth it if MLS comes in soon.