adding a timeout to stream metadata health checks to increase failure visibility (#1251)
in the event of a hanging request, the health check failure is not
registered by the stream metadata service, but it is registered by the
load balancer. this causes the service to get terminated, but the
service is unable to detect the failure and log the error. this PR
allows us to capture and log the error so that we can respond to it
better