Annotate returned server errors with node address and url (#1261)
For
https://linear.app/hnt-labs/issue/TOWNS-12853/[omega][user-feedback]-error-creating-town,
we were uncertain if this was a service error or an error with rpc
provider. This PR annotates river errors with node metadata so that when
we view harmony console logs in datadog, we can correlate the errors
manually and determine if error patterns are only appearing on certain
nodes.