adding cache-control headers to response logs in stream metadata (#1275)
we're seeing multiple requests hitting the same `/media` endpoint, which
should never happen. and there are a few other routes that are supposed
to be cached and should not make it to the server, but we're still
seeing them. i'm adding cache-control headers to the logger just to see
if there's a pattern