clientDecryptionExtensions, bugFix/perf improvement
I have more and larger refactors coming, but this seemed useful to call out
participants includes left users, and instantiates a new set every time with joined, invited, and left users. And it is called inside a pretty hot loop.
I will optimize the hot loop in an upcoming pr