stream-metadata: add support for `/refreshStatus/:invalidationId` (#1351)
We need to wait for refresh to finish before emitting `refreshMetadata`
for space members nft.
Previously we tried polling both space and space member metadata and
waiting eventId of the image to be equal
The previous attempt didn’t worked and also increased the load on the
service.

This PR solves it, adding a `/refreshStatus/:invalidationId` route. 
Now after a refresh we get the invalidationId, which we can use to poll
`/refreshStatus/:invalidationId` on the client until the refresh is
completed.

We dont need `/event` in the space image route anymore. I'll create a PR
later