Another attempt at fixing the flaky waitFor timeout waiting for streamMembershipUpdated (#1443)
These are always failing when they go through this utility

Our sync startup code is wild. Really hard to trace it all. My gut tells
me there is a race where we either subscribe to a stream too soon or too
late and we don’t get the push that the membership event was added to
the space stream, but I can’t figure out where it is. If this fixes it
we should burn all that code down and make something simpler.