[RIVER-1305] Migrate db schema to use fixed # of partitions (#1349)
- Migrate schema from strategy to create tables for each stream to a
static allocation of partitions, supporting a partition count between 1
and 256 inclusive.
- the partition is determined by the lowest byte of the xxhash64 of the
stream id string after we take the modulus against the partition count.
- test cases do not run with the full set of 256 partitions, which takes
~2s on my local machine to allocate per test case, and significantly
slows down unit tests. Instead, we use 4 partitions in testing. This is
supported with custom migration scripts and a TestMode flag in the db
config.

There is config on the node to determine if the pg store will use the
legacy or migrated schema for storing stream data.

Regular store tests cases have been migrated to use the most both
schemas. In addition there is a new test file that tests unmigrated
streams to behave correctly after a store changes to use the migrated
schema.

E2E tests and rpc tests are still using legacy schema - I left it this
way because we made such significant changes to existing logic that I'd
like CI tests to run against these for a few days before we switch over.

Follow-up work: detect if the entire stream corpus is migrated on
startup and short-circuit the migration check for each stream before
data access.