Capture membership evaluation errors when no wallets evaluate as space members (#1495)
When multiple linked wallets are checked for space membership and none
of them are conclusively determined to be members, we ignore any errors
in invoked checks and return false. However, if any wallet check failed,
we cannot conclusively determine that the user was not a space member
and should return an entitlement evaluation error so that the user will
not be considered unentitled by the caller.

This PR captures all errors generated when conducting membership checks
and returns an aggregate error in the case that all checks evaluated and
none of them were positive. The user is consider "not a member" only if
there are no errors whatsoever.

#1494

---------

Co-authored-by: Serge Khorun <104387024+sergekh2@users.noreply.github.com>
Co-authored-by: Serge Khorun <serge@hntlabs.com>