Don’t allow more than one simultaneous stream/media/client request in stream-metadata service
I think this is a first step before adding a LRU cache outlined here: https://github.com/river-build/river/pull/1345

This service is super bursty, if we have lots of traffic coming into it hopefully this will serve everything with the same resources.

Decrypting the space image is already debounced at the StreamStateView level, so this should make a visible impact.