NET-142 dropfacet shim (#1556)
NET-142 - We need to expose the IDropFacet interface for river.build
drop page. This PR wraps the IDropShim with the current abi in a
convenience factory function to instantiate the class. Everything in
this PR will be consumable in river.build NextJS app through the web3
npm package that will ship with this once this Pr is merged.