stop current decryption extensions in “resetCrypto” function (#1588)
I guess I forgot that this function was here, because I’m not sure who
else would have written it, but apparently we have a test that calls
resetCrypo in a loop, and I noticed in the new vite tests its throwing
warnings about 10+ listeners added to the same event
https://github.com/river-build/river/actions/runs/11942779116/job/33290511830?pr=1582

this pr fixes this issue. This function isn’t called anywhere but the
tests.