notifications: delete expired webpush subscription automatically (#1629)
When the notification service tries to send a webpush notification and
receives a `410 - Gone` the subscription is expired. This indicates that
either:

- the user unsubscribed from notifications in the browser
- the browser has renewed the subscription and the notification service
didn't got an update with the new endpoint.

This change deletes the expired subscription from the database. It takes
the endpoint into considering meaning that in case the user refreshed
the subscription with the new endpoint there is no race condition that
deletes it immediately due to the 410 error.