Update space membership join to use isSpaceMember (#1644)
#1643 

This change updates the entitlement check for a adding a user to the
space stream to check for space membership instead of looking for the
READ permission. Since the requested auth was previously of kind
`chainAuthKindSpace`, previously the authorization module would look for
at least 1 role on the space with READ permission that the wallet
satisfied in order to allow the join event, which makes no sense.
However, we ended up satisfying this in practice because the member role
is typically an EVERYONE User Entitlement role, and it has the READ
permission attached.

The new check is a a chainAuthArgs of kind `chainAuthKindIsSpaceMember`,
which does not check any entitlements at all, it just validates that the
user is a space member.

This permissions check actually relaxes the strictness of checks here
because all auth checks already include a space membership check - so we
were already preventing all non-members from adding themselves to the
space stream. We just don't also include the erroneous check for a READ
entitlement anymore. The system behavior likely will not change as a
result of this change unless a space owner or other admin modifies the
member role.