Stream and xchain services to share Blockchains (#1661)
#1659 

Refactor stream and xchain services to use the same base chain. Comments
included in the code for clarity.

I also saw some code cruft around unused variables in the transaction
pool and went ahead and cleaned it up. Third small change - swapped in a
utility function in the stream node for determining wallet address since
it was line-for-line duplicated code.