Rpc Connection Refactor/Deduplication (#1721)
We started copying rpc connections everywhere because some needed
connect-node and that doesn’t work if you import it in a browser. I just
did the inline import thing and now we can run everywhere/anywhere.

Also worth noting, the new delta nodes only accept http2 connections so
without this you have to jump through the goofy hoops to get it to work.