SpaceDapp: change tip to receive tokenId instead of receiver address (#1819)
I thought it made more sense to be able to pass receiver here. But after
writing some tests, I realized the membership nft could be in any linked
wallet - and that is where the tip is going to go.

It's confusing to have a parameter called receiver, then internally
fetch linked wallets, and have the tip go potentially to another
address. So instead callers can pass the tokenId using getTokenIdOfOwner