Tip events, properly handle the to user address (#1903)
The sender and receiver of the tip are the wallets that hold the space
nfts, not the root wallet addresses. So we can’t assume that the
receiver will have a user stream. We need to specify the to user address
in the metadata and verify on appending a received transaction that this
actually belongs to the user.