Add logging metadata to help debug race conditions in replication (#1893)
This PR separates out some additional logging and population of error
metadata in returned errors that mostly only occurs in the case of error
conditions, and should not affect total log volume overly. The
additional metadata has been helpful in tracking down what may be
happening when searching for race conditions in the code.

It also adds a service tester configuration to enable logging for
specific test cases.