[RIVER-1858] Node and corrupt stream metrics (#1857)
#1858 

Consider a stream corrupt if it was unavailable for greater than 50
consecutive update attempts. The archiver rotates between nodes whenever
an update fails, so in the future when replication is deployed we can
reduce this number to something like 5 since at that point we will have
tried every node. The 50 is in there to accomodate stream unavailability
while nodes are updating.

Once a stream is corrupt, it will not attempt to retry updating itself.
New miniblocks on the network will trigger the stream, however, to
update itself... but it will no longer retry on failures. This is an
in-memory setting and will be reset the next time the node starts.