WIP: Go MLS snapshotting (Help wanted) (#1963)
I could use some help here but this PR more or less describes what we
need to accomplish.

Some parts of MLS are impossible for Go to snapshot on its own since it
doesn't understand the commits — instead we need to round up all commits
and process them in the Rust MLS Snapshot code, while keeping track of
the latest valid `GroupInfoMessage`.

There is no way to _guarantee_ that an invalid commit slips through our
checks and makes it into the stream, this is why we cannot simply rely
on the latest `groupInfoMessage`, we need to rely on the latest
`groupInfoMessage` accompanied by a valid commit. There are more issues
to solve here but this is a start.

I wrote some comments inline, ideas more than welcome!

stacked on #1962