notification service: bugfix tracking membership in new towns (#1993)
Currently the notification service doesn't handle sync updates that
include miniblocks. The result is that once events are dropped from the
minipool calculating the stream membership must be done from `snapshot +
blocks + minipool` but is done from `snapshot + minipool`.

For notifications this is not an issue because these are handled
immediately when the notification service receives them through a sync
update. For user membership it needs to respect join/leave events in
miniblocks.

This PR adds an unittest that hits the problem and introduces the
`ApplyBlock` function that is called for incoming miniblocks.