[RIVER-1859] Miniblock scrubber (#1943)
#1859 

Built a miniblock scrubber that checks for corrupted blocks in the
stream. As part of this work I expanded the existing miniblockInfo
construction to optionally validate a bunch of extra stuff. A piece of
follow-up work here is validating snapshot state based on the miniblocks
that have come in since the previous snapshot.

Design
======
- the miniblock scrubber will always schedule a scrub, it does not deny
a scheduled scrub based on the recency of the last scrub
- The scrubber scrubs from a minimum block to the latest block, and
outputs a report to a channel
- The scrubber expects a report consumer to be constantly polling, it
will block the scrubber if the buffer becomes full because tasks block
on emitting to the channel

Unit tests
=======
- a single-block stream
- a multi-block stream
- a corrupted stream

Extra stuff:
========
- rename existing scrubber for easier disambiguation of scrubbers
- some spelling fixes