keep track of welcome messages inside snapshot (#2005)
This is a simpler approach than #2004 

What we need is a foolproof way to keep track of where welcome messages
are stored, since clients need them to join the group. We also want to
avoid blind scrollbacks since that will be too inefficient and is
non-deterministic (what if the welcome pack isn't in there?)

This PR:

When new users join an existing stream, they will upload their
`KeyPackages`.
Existing users will process the `KeyPackages`, run
`Group.addMembers(keyPackages)` and publish an event containing:
```
{
  commit: []byte // commit mutating the group state
  group_info_message: []byte // new updated group info msg for future members
  signature_public_keys: [][]byte // which sigs are included by this op
  welcome_messages: [][]byte // welcome msgs that can be used by the new users to bootstrap their group
}
```

this PR makes sure that returning clients:
- know that they've been sent a welcome message
- know where to find it (miniblock num)