notification service: new APN notification payload (#1994)
Change APN payload format to reduce its size.

Strip the stream event before the notification is send if the size is
guaranteed too large. If APN still reports that the payload is too large
retry it with the stream event stripped from it.

Merge when there is clarity if we need to be backwards compatiblity.