unittest: fix flaky notifications test (#2053)
This change distinguishes between streams that the notification service
will track between streams that already exists for a while and streams
that are just allocated. For just allocated streams the notification
service will process events in miniblocks/minipool that it got in the
sync reset when subscribing. For streams that already exists it will
ignore these to prevent sending notifications multiple times for the
same stream.