Update Xchain to use entitlement coordinator, establish tests for using V1 and V2 check events (#2151)
Branched from G's entitlement coordinator PR here, this PR updates the
xchain service to respond to both v1 and v2 checks. The mock entitlement
checker is also updated to include xchain methods so the final event
will be emitted and the client simulator library can detect it and
validate that the check reached quorum.