[BUGFIX] Error handling fixes in pg stream store, notifications store (#2185)
As part of investigating the pg code for archiver go test flakes, I
noticed and fixed the following:
- whenever we iterate over a result set using `rows.Next()` and
`rows.Next()` returns false, according to the pgx documentation we are
supposed to check for an error on `rows.Err()` before assuming we
correctly ran through the entire result set, but we weren't doing this
in several places. So, I swapped the file's predominant row iteration
pattern with a pattern that uses the `pgx.ForEachRow` function, which
automatically closes the result set stream and checks for errors during
and after row scans.
- In some places in the code we were checking to see if `err ==
context.Canceled` or `err == pgx.*` instead of checking it's unwrap tree
with `errors.Is` - this was unlikely to produce error behaviors based on
where we used it, but I fixed anyway since it's not a good pattern.

I also made a lot of more general `if err = ...; err != nil` changes for
the sake of concision.

Why did this come up? There is an occasional test flake that comes up
with the archiver where I added some extra debug information and noticed
that the transaction is aborted when we get to the error check that
fails the test case, but the last `Query` call succeeded. If there was
an issue in the pg streaming connection for reading rows that lead this
transaction to fail, then we'd like to capture that so we can be aware
of it and have more information about how to debug this flake.