node: set a fixed miner tip for river chain transactions (#2218)
For each transaction that is assembled through the generated abi
bindings the node performs 2 rpc calls to determine the miner tip. At
the moment there is no gas market on River chain and no need to ask the
a suggested miner tip fee.

This PR sets a hard coded miner tip on River chain transactions. We can
also make it an on-chain config setting but I don't think that is needed
at the moment.