[BUGFIX] - Fix test schema cleanup and enforce proper cleanup by test writers (#2220)
This is a low priority bugfix of test case cleanup. Sometimes schemas
were not being deleted because the context used for schema deletion
would be cancelled since it was derived from a context created for the
lifetime of a particular test object, and that object was cleaned up
before the schema was deleted. For simplicity of reasoning, I swapped
the argument ctx with context.Background() and added a panic to ensure
that schemas are always properly deleted.