Cover edge case where we lose sync and recover and re-initialize the user inbox stream (#2267)
if lots of keys have come in in the meantime we will not download them
until the user refreshes the page. If a key comes in via sync, we would
reset the indexes in the stream and never download them.

This is probably the long standing bug where some random message doesn’t
sync randomly.