Refactor notification stream syncing code for tandem use by bot service (#2280)
The actual streams tracker will be sufficiently different for each
service that it makes sense to go ahead and implement them separately:
the notification service cares about all DM, GDM, and channel streams,
and tracks all of these, but the bot registry service only cares about
streams associated with a registered bot, and the channels in the spaces
that bot belongs to. Furthermore, channel streams are sometimes picked
up on allocation by the notification service, but they are almost always
going to be added retroactively for the bot registry service at the
point at which we detect a bot has been added to a channel.

However, the logic for keeping an up-to-date internal copy of the view
of each stream, and for syncing that stream, are very similar. This PR
refactors those common elements into a shared library and plugs them
back into the notification service.