Refactor stream syncing and add to bot registry service (#2347)
After discussing requirements with Austin, we decided it's fine for now
for the bot registry to over-subscribe to irrelevant streams for the
sake of simplicity, meaning it can share almost all of the logic the
notifications service uses for syncing and tracking streams.

This PR refactors the stream tracker a bit to share it between the
notifications service and the bot registry service, and adds rudimentary
stream tracking to the bot registry service. There is also a simple test
case that confirms channel events are being forwarded to the listener,
which will eventually be responsible for determining which events need
to be sent to which webhook, queueing those events, and provoking the
bot service to solicit keys.