Add the web3 folder to the generate_sdk_index script (#2394)
I think we didn’t do this because we wanted to hide the implementation
because we wanted to be able to swap out the implementation of the smart
contracts…
but we don’t want to do that any more.
this makes it much easier to keep everything up to date
not to mention modular and reusable.
wondering if this breaks anything?